Updated French translations (and removed pro features)
I completed translations so **everything**, 100% is in french (except things that are sadly marked as non-translatable by upstream) so there is no missing translation by now (although, GitHub.com is not translated in French).
I also chose to remove strings that refer to the upstream premium features, and ads, and donation system. **If that's a problem, tell me and I'll add these strings back.**
I didn't try the app with these new translations, although it should work.